backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
playwright==1.40.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
transformers==4.35.2
torch==2.1.1